Add support for new query APIs. (#23)
This adds support for the new asynchronous query APIs. The cda-python wrapper hides the async queries so there are (almost) no changes to the python API.

New fields in `Result`:
- `Result.total_row_count` is the total number of rows in the query response
- `Result.has_next_page` is true if there are more rows of data

The API now makes a distinction between `limit` and `page_size`. `limit` is the limit of total rows in the query response. If this is not provided, there is no limit and all rows will be generated. `page_size` is the number of rows in a page of data. This is the one API change which could affect users. In most cases users will want to omit `limit` and use `page_size` where they used to use `limit`.

Requires cda-client v1.1.0

Fixes #11